Refactor: 설정초기화
App.js -> App.tsx
