+++ frontend/src/App.css
@@ -1,11 +1,11 @@
 body{
-  margin : 0 ;
-  padding: 0 ;
-}
-.container{
-  display: flex;
-  justify-content: center; /* 수평 가운데 정렬 */
-  align-items: center; /* 수직 가운데 정렬 */
-  height: 100%;
-  margin: 0 auto;
-}
+    margin : 0 ;
+    padding: 0 ;
+  }
+  .container{
+    display: flex;
+    justify-content: center; /* 수평 가운데 정렬 */
+    align-items: center; /* 수직 가운데 정렬 */
+    height: 100%;
+    margin: 0 auto;
+  }
