+++ frontend/src/Components/Nav/Nav.tsx
@@ -2,7 +2,9 @@ import React from 'react'
 
 function Nav() {
   return (
-    <div>Nav</div>
+    <div>
+      Nav
+    </div>
   )
 }
