+++ frontend/src/Pages/Join/Index.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Join() {
+  return (
+    <div>
+      join
+    </div>
+  )
+}
+
+export default Join
