+++ frontend/src/Pages/Main/Index.tsx
@@ -2,7 +2,9 @@ import React from 'react'
 
 function Main() {
   return (
-    <div>Main</div>
+    <div>
+      Main
+    </div>
   )
 }
