+++ frontend/src/Pages/Meet/Index.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Meet() {
+  return (
+    <div>
+      미트
+    </div>
+  )
+}
+
+export default Meet
