--- frontend/src/pages/login/Login.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function login() {
-  return (
-    <div>login</div>
-  )
-}
-
-export default login;
