--- frontend/src/pages/notice/Notice.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Notice() {
-  return (
-    <div>Notice</div>
-  )
-}
-
-export default Notice
