--- frontend/src/pages/opens/Meetings.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function Meetings() {
-  return (
-    <div>Meetings</div>
-  )
-}
-
-export default Meetings
